Don't swallow errors logged at the top level
Previously errors were expected to be logged as e.g.

    logger.warn('problem', {error: new MyError('foo')})

However in some cases we were logging errors as e.g.

    logger.warn('problem', new MyError('foo'))

In the latter case, nothing about the error was getting written to the log
file. Check for this top-level error case.

I'm not sure the behavior is appropriate - please double check that the test
does the right thing.